Fix conversions

+++ src/itdelatrisu/opsu/states/Game.java
@@ -814,7 +814,7 @@ public class Game extends BasicGameState {
 				GameImage.INPUTOVERLAY_KEY.getImage().getScaledCopy((int) BTNSIZE, (int) BTNSIZE);
 			for (int i = 0; i < 4; i++) {
 				inputOverlayKeys[i].render(g, x, y, keyimg);
-				y += BTNSIZE;
+				y += (int) BTNSIZE;
 			}
 		}
 
@@ -2414,7 +2414,7 @@ public class Game extends BasicGameState {
 						progress = (float) (trackPosition - breakTime) / approachTime;
 					else if (endTime - trackPosition < approachTime)
 						progress = (float) (endTime - trackPosition) / approachTime;
-					flashlightRadius += (width - flashlightRadius) * progress;
+					flashlightRadius += (width - flashlightRadius) * ((int) progress);
 				}
 			} else if (flashlightRadius != targetRadius) {
 				// radius size change
@@ -2424,7 +2424,7 @@ public class Game extends BasicGameState {
 					if (flashlightRadius < targetRadius)
 						flashlightRadius = targetRadius;
 				} else {
-					flashlightRadius += radiusDiff;
+					flashlightRadius += (int) radiusDiff;
 					if (flashlightRadius > targetRadius)
 						flashlightRadius = targetRadius;
 				}
